requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
payos>=1.0.0
python-dotenv>=1.0.0
Django>=6.0.0
djangorestframework>=3.15.0
PyMySQL>=1.1.1
django-cors-headers>=4.3.0